requirements.txt
annotated-types==0.7.0
nginxfmt==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
